Finish first iteration
